search_api_solr/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-json-logger>=2.0.7
prometheus-fastapi-instrumentator>=6.1.0
redis>=4.5.0
aioredis>=2.0.1
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
